+++ src/types/data.ts
@@ -1,0 +1,108 @@
+export interface User {
+  id: number;
+  username: string;
+  image: string;
+  bio?: string;
+
+  // SocialUserData fragment
+  url?: string;
+  socials?: Array<UserSocial>;
+  firstName?: string;
+  lastName?: string;
+  displayName?: string;
+  fullName?: string;
+  followCount?: number;
+  followerCount?: number;
+
+  // PlanUserData fragment
+  isUserHacker?: boolean;
+  isUserPro?: boolean;
+
+  // RolesUserData fragment
+  roles?: Array<UserRole>;
+}
+
+export interface UserSocial {
+  id: number;
+  url: string;
+  type: UserSocialType;
+}
+export enum UserSocialType {
+  twitter = "twitter",
+  github = "github",
+  linkedin = "linkedin",
+  website = "website",
+  youtube = "youtube",
+  twitch = "twitch",
+  facebook = "facebook",
+  discord = "discord",
+}
+export interface UserRole {
+  id: number;
+  name: string;
+  key: string;
+  tagline: string;
+}
+
+// Repl
+export interface Repl {
+  id: string;
+  url: string;
+  title: string;
+  description: string;
+  timeCreated: string;
+  slug: string;
+  isPrivate: boolean;
+
+  // SocialReplData fragment
+  likeCount?: number;
+  publicForkCount?: number;
+  runCount?: number;
+  commentCount?: number;
+  tags?: Array<Tag>;
+  iconUrl?: string;
+  imageUrl?: string;
+
+  // CommentsReplData fragment
+  comments?: {
+    items?: Array<ReplComment>;
+  };
+
+  // OwnerData fragment
+  owner?: Pick<User, "id" | "username" | "image"> & {
+    __typename: string;
+    description?: string;
+  };
+
+  // MultiplayersData fragment
+  multiplayers?: Array<User>;
+}
+
+export interface Tag {
+  id: string;
+  isOfficial: boolean;
+}
+export interface ReplComment {
+  id: number;
+  body: string;
+  user: User;
+}
+
+export interface UserDataInclusion {
+  includeSocialData?: boolean;
+  includeRoles?: boolean;
+  includePlan?: boolean;
+}
+
+export interface ReplDataInclusion {
+  includeSocialData?: boolean;
+  includeComments?: boolean;
+  includeOwner?: boolean;
+  includeMultiplayers?: boolean;
+}
+
+export type GraphResponse<T> = Promise<T | never>;
+
+export type ReplQueryOutput = GraphResponse<{ repl: Repl }>;
+export type UserByUsernameQueryOutput = GraphResponse<{ userByUsername: User }>;
+export type UserQueryOutput = GraphResponse<{ user: User }>;
