+++ src/types/index.ts
@@ -1,423 +1,33 @@
-/*****************************************************************
- * * FS (Filesystem) Types
- *****************************************************************/
+import {
+  DirectoryChildNode,
+  WatchFileListeners,
+  WatchTextFileListeners,
+  WatchDirListeners,
+} from "./fs";
+import {
+  UserDataInclusion,
+  UserQueryOutput,
+  UserByUsernameQueryOutput,
+  ReplDataInclusion,
+  ReplQueryOutput,
+} from "./data";
+import {
+  ThemeValuesGlobal,
+  ThemeVersion,
+  OnThemeChangeValuesListener,
+} from "./themes";
+import { OnActiveFileChangeListener } from "./session";
 
-/**
- * A Filesystem node type
- */
-export enum FsNodeType {
-  File = "FILE",
-  Directory = "DIRECTORY",
-}
-
-/**
- * A base interface for nodes, just includes
- * the type of the node and the path, This interface
- * does not expose the node's content/children
- */
-export interface FsNode {
-  path: string;
-  type: FsNodeType;
-}
-
-export enum ChangeEventType {
-  Create = "CREATE",
-  Move = "MOVE",
-  Delete = "DELETE",
-  Modify = "MODIFY",
-}
-
-export interface MoveEvent {
-  eventType: ChangeEventType.Move;
-  node: FsNode;
-  to: string;
-}
-
-export interface DeleteEvent {
-  eventType: ChangeEventType.Delete;
-  node: FsNode;
-}
-
-export interface WatchFileListeners<T extends string | Blob = string> {
-  onChange: WatchFileOnChangeListener<T>;
-  onError?: WatchFileOnErrorListener;
-  onMoveOrDelete?: WatchFileOnMoveOrDeleteListener;
-}
-
-export interface WatchTextFileListeners {
-  onReady: WatchTextFileOnReadyListener;
-  onChange?: WatchTextFileOnChangeListener;
-  onError?: WatchTextFileOnErrorListener;
-  onMoveOrDelete?: WatchTextFileOnMoveOrDeleteListener;
-}
-
-export interface DirectoryChildNode {
-  filename: string;
-  type: FsNodeType;
-}
-
-/*****************************************************************
- * * Theme Types
- *****************************************************************/
-export type CssColor = string;
-
-export interface ThemeValuesGlobal {
-  __typename?: string;
-  backgroundRoot: CssColor;
-  backgroundDefault: CssColor;
-  backgroundHigher: CssColor;
-  backgroundHighest: CssColor;
-  backgroundOverlay: CssColor;
-  foregroundDefault: CssColor;
-  foregroundDimmer: CssColor;
-  foregroundDimmest: CssColor;
-  outlineDimmest: CssColor;
-  outlineDimmer: CssColor;
-  outlineDefault: CssColor;
-  outlineStronger: CssColor;
-  outlineStrongest: CssColor;
-  accentPrimaryDimmest: CssColor;
-  accentPrimaryDimmer: CssColor;
-  accentPrimaryDefault: CssColor;
-  accentPrimaryStronger: CssColor;
-  accentPrimaryStrongest: CssColor;
-  accentPositiveDimmest: CssColor;
-  accentPositiveDimmer: CssColor;
-  accentPositiveDefault: CssColor;
-  accentPositiveStronger: CssColor;
-  accentPositiveStrongest: CssColor;
-  accentNegativeDimmest: CssColor;
-  accentNegativeDimmer: CssColor;
-  accentNegativeDefault: CssColor;
-  accentNegativeStronger: CssColor;
-  accentNegativeStrongest: CssColor;
-  redDimmest: CssColor;
-  redDimmer: CssColor;
-  redDefault: CssColor;
-  redStronger: CssColor;
-  redStrongest: CssColor;
-  orangeDimmest: CssColor;
-  orangeDimmer: CssColor;
-  orangeDefault: CssColor;
-  orangeStronger: CssColor;
-  orangeStrongest: CssColor;
-  yellowDimmest: CssColor;
-  yellowDimmer: CssColor;
-  yellowDefault: CssColor;
-  yellowStronger: CssColor;
-  yellowStrongest: CssColor;
-  limeDimmest: CssColor;
-  limeDimmer: CssColor;
-  limeDefault: CssColor;
-  limeStronger: CssColor;
-  limeStrongest: CssColor;
-  greenDimmest: CssColor;
-  greenDimmer: CssColor;
-  greenDefault: CssColor;
-  greenStronger: CssColor;
-  greenStrongest: CssColor;
-  tealDimmest: CssColor;
-  tealDimmer: CssColor;
-  tealDefault: CssColor;
-  tealStronger: CssColor;
-  tealStrongest: CssColor;
-  blueDimmest: CssColor;
-  blueDimmer: CssColor;
-  blueDefault: CssColor;
-  blueStronger: CssColor;
-  blueStrongest: CssColor;
-  blurpleDimmest: CssColor;
-  blurpleDimmer: CssColor;
-  blurpleDefault: CssColor;
-  blurpleStronger: CssColor;
-  blurpleStrongest: CssColor;
-  purpleDimmest: CssColor;
-  purpleDimmer: CssColor;
-  purpleDefault: CssColor;
-  purpleStronger: CssColor;
-  purpleStrongest: CssColor;
-  magentaDimmest: CssColor;
-  magentaDimmer: CssColor;
-  magentaDefault: CssColor;
-  magentaStronger: CssColor;
-  magentaStrongest: CssColor;
-  pinkDimmest: CssColor;
-  pinkDimmer: CssColor;
-  pinkDefault: CssColor;
-  pinkStronger: CssColor;
-  pinkStrongest: CssColor;
-  greyDimmest: CssColor;
-  greyDimmer: CssColor;
-  greyDefault: CssColor;
-  greyStronger: CssColor;
-  greyStrongest: CssColor;
-  brownDimmest: CssColor;
-  brownDimmer: CssColor;
-  brownDefault: CssColor;
-  brownStronger: CssColor;
-  brownStrongest: CssColor;
-  black: CssColor;
-  white: CssColor;
-}
-
-export enum ColorScheme {
-  Light = "light",
-  Dark = "dark",
-}
-
-export interface CustomTheme {
-  author: User;
-  colorScheme: ColorScheme;
-  hasUnpublishedChanges: boolean;
-  id: number;
-  isCurrentUserThemeAuthor: boolean;
-  isInstalledByCurrentUser: boolean;
-  latestThemeVersion: ThemeVersion;
-  numInstalls?: number;
-  slug?: string;
-  status?: "public" | "private";
-  title?: string;
-}
-
-export interface ThemeSyntaxHighlightingTag {
-  __typename: string;
-  name: string;
-  modifiers: null | Array<string>;
-}
-
-export interface ThemeSyntaxHighlightingModifier {
-  textDecoration?: string;
-  fontSize?: string;
-  fontWeight?: string;
-  fontStyle?: string;
-  color?: string;
-}
-
-export interface ThemeEditorSyntaxHighlighting {
-  __typename: string;
-  tags: Array<ThemeSyntaxHighlightingTag>;
-  values: ThemeSyntaxHighlightingModifier;
-}
-
-export interface ThemeValuesEditor {
-  editor: Array<ThemeEditorSyntaxHighlighting>;
-}
-
-export interface ThemeValues {
-  __typename?: string;
-  editor: ThemeValuesEditor;
-  global: ThemeValuesGlobal;
-}
-
-export interface ThemeVersion {
-  __typename?: string;
-  id: number;
-  hue: number;
-  lightness: number;
-  saturation: number;
-  timeUpdated?: string;
-  description?: string;
-  customTheme?: CustomTheme;
-  values?: ThemeValues;
-}
-
-/*****************************************************************
- * * Data Input/Output Types
- *****************************************************************/
-
-export interface UserDataInclusion {
-  includeSocialData?: boolean;
-  includeRoles?: boolean;
-  includePlan?: boolean;
-}
-
-export interface ReplDataInclusion {
-  includeSocialData?: boolean;
-  includeComments?: boolean;
-  includeOwner?: boolean;
-  includeMultiplayers?: boolean;
-}
-
-export type GraphResponse<T> = Promise<T | never>;
-
-export type ReplQueryOutput = GraphResponse<{ repl: Repl }>;
-export type UserByUsernameQueryOutput = GraphResponse<{ userByUsername: User }>;
-export type UserQueryOutput = GraphResponse<{ user: User }>;
-
-/*****************************************************************
- * * GraphQL Types
- *****************************************************************/
-
-// User
-export interface User {
-  id: number;
-  username: string;
-  image: string;
-  bio?: string;
-
-  // SocialUserData fragment
-  url?: string;
-  socials?: Array<UserSocial>;
-  firstName?: string;
-  lastName?: string;
-  displayName?: string;
-  fullName?: string;
-  followCount?: number;
-  followerCount?: number;
-
-  // PlanUserData fragment
-  isUserHacker?: boolean;
-  isUserPro?: boolean;
-
-  // RolesUserData fragment
-  roles?: Array<UserRole>;
-}
-
-export interface UserSocial {
-  id: number;
-  url: string;
-  type: UserSocialType;
-}
-export enum UserSocialType {
-  twitter = "twitter",
-  github = "github",
-  linkedin = "linkedin",
-  website = "website",
-  youtube = "youtube",
-  twitch = "twitch",
-  facebook = "facebook",
-  discord = "discord",
-}
-export interface UserRole {
-  id: number;
-  name: string;
-  key: string;
-  tagline: string;
-}
-
-// Repl
-export interface Repl {
-  id: string;
-  url: string;
-  title: string;
-  description: string;
-  timeCreated: string;
-  slug: string;
-  isPrivate: boolean;
-
-  // SocialReplData fragment
-  likeCount?: number;
-  publicForkCount?: number;
-  runCount?: number;
-  commentCount?: number;
-  tags?: Array<Tag>;
-  iconUrl?: string;
-  imageUrl?: string;
-
-  // CommentsReplData fragment
-  comments?: {
-    items?: Array<ReplComment>;
-  };
-
-  // OwnerData fragment
-  owner?: Pick<User, "id" | "username" | "image"> & {
-    __typename: string;
-    description?: string;
-  };
-
-  // MultiplayersData fragment
-  multiplayers?: Array<User>;
-}
-
-export interface Tag {
-  id: string;
-  isOfficial: boolean;
-}
-export interface ReplComment {
-  id: number;
-  body: string;
-  user: User;
-}
-
-/*****************************************************************
- * * Miscalleneous / React Types / Function Args
- *****************************************************************/
-
-export enum UseWatchTextFileStatus {
-  Error = "error",
-  Loading = "loading",
-  Watching = "watching",
-  Moved = "moved",
-  Deleted = "deleted",
-}
+export * from "./fs";
+export * from "./themes";
+export * from "./data";
+export * from "./session";
 
 export enum HandshakeStatus {
   Ready = "ready",
   Error = "error",
   Loading = "loading",
 }
-
-export interface StrError {
-  error: string;
-}
-export interface NullableStrError {
-  error: string | null;
-}
-
-export interface TextChange {
-  from: number;
-  to?: number;
-  insert?: string;
-}
-
-export type WriteChange = (changes: TextChange | Array<TextChange>) => void;
-export type GetLatestContent = () => string;
-
-export interface TextFileReadyEvent {
-  writeChange: WriteChange;
-  getLatestContent: GetLatestContent;
-  initialContent: string;
-}
-export interface TextFileOnChangeEvent {
-  changes: Array<TextChange>;
-  latestContent: string;
-}
-
-export interface WatchDirListeners {
-  onChange: WatchDirOnChangeListener;
-  onMoveOrDelete?: WatchDirOnMoveOrDeleteListener;
-  onError: WatchDirOnErrorListener;
-}
-
-export type WatchDirOnErrorListener = (
-  err: Error,
-  extraInfo?: Record<string, any>
-) => void;
-export type WatchDirOnChangeListener = (children: Array<FsNode>) => void;
-export type WatchDirOnMoveOrDeleteListener = (
-  event: DeleteEvent | MoveEvent
-) => void;
-export type OnActiveFileChangeListener = (file: string) => void;
-export type WatchFileOnChangeListener<T extends string | Blob = string> = (
-  newContent: T
-) => void;
-export type WatchFileOnErrorListener = (error: string) => void;
-export type WatchFileOnMoveOrDeleteListener = (
-  moveOrDeleteEvent: MoveEvent | DeleteEvent
-) => void;
-export type WatchTextFileOnReadyListener = (
-  readyEvent: TextFileReadyEvent
-) => void;
-export type WatchTextFileOnChangeListener = (
-  changeEvent: TextFileOnChangeEvent
-) => void;
-export type WatchTextFileOnErrorListener = (error: string) => void;
-export type WatchTextFileOnMoveOrDeleteListener = (
-  moveOrDeleteEvent: MoveEvent | DeleteEvent
-) => void;
-export type HandshakeOuput = Promise<null | (() => void)>;
-export type OnThemeChangeValuesListener = (values: ThemeValuesGlobal) => void;
-export type OnThemeChangeListener = (theme: ThemeVersion) => void;
 
 export interface ReplitInitArgs {
   timeout?: number;
@@ -427,6 +37,15 @@ export interface ReplitInitArgs {
 export interface ReplitInitOutput {
   dispose: () => void;
   status: HandshakeStatus;
+}
+
+export type HandshakeOuput = Promise<null | (() => void)>;
+
+export interface StrError {
+  error: string;
+}
+export interface NullableStrError {
+  error: string | null;
 }
 
 /*****************************************************************
