+++ src/types/session.ts
@@ -1,0 +1,1 @@
+export type OnActiveFileChangeListener = (file: string) => void;
